services/payment/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary>=2.9.10
pydantic==2.5.3
python-dotenv==1.0.0
stripe==11.1.0
requests==2.31.0
httpx==0.28.1